docker and error codes

+++ Cryptocop.Software.API/Cryptocop.Software.API/Controllers/AddressController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Cryptocop.Software.API.Services.Interfaces;
 using Cryptocop.Software.API.Models.InputModels;
-
+using Cryptocop.Software.API.Models.Exceptions;
 namespace Cryptocop.Software.API.Controllers
 {
     [Authorize]
@@ -22,12 +22,18 @@ namespace Cryptocop.Software.API.Controllers
         //Gets all addresses associated with authenticated user
         public IActionResult GetAllAddresses()
         {
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             return Ok(_addressService.GetAllAddresses(User.Identity.Name));
         }
         [HttpPost]
         [Route("")]
         //Adds a new address associated with authenticated user, see
         public IActionResult AddAddress([FromBody] AddressInputModel address){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             _addressService.AddAddress(User.Identity.Name, address);
             return NoContent();
         }
@@ -35,6 +41,9 @@ namespace Cryptocop.Software.API.Controllers
         [HttpDelete]
         [Route("{id}")]
         public IActionResult DeleteAddress(int id){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             _addressService.DeleteAddress(User.Identity.Name, id);
             return NoContent();
         }
