+++ Cryptocop.Software.API/Cryptocop.Software.API/Controllers/ShoppingCartController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Cryptocop.Software.API.Services.Interfaces;
 using Cryptocop.Software.API.Models.InputModels;
+using Cryptocop.Software.API.Models.Exceptions;
 
 namespace Cryptocop.Software.API.Controllers
 {
@@ -21,6 +22,9 @@ namespace Cryptocop.Software.API.Controllers
         [Route("")]
         //Gets all items within the shopping cart for current user
         public IActionResult GetCartItems(){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             return Ok(_shoppingcartService.GetCartItems(User.Identity.Name));
         }
 
@@ -28,6 +32,9 @@ namespace Cryptocop.Software.API.Controllers
         [Route("")]
         //Adds an item to the shopping cart for current user
         public async Task<IActionResult> AddCartItem([FromBody] ShoppingCartItemInputModel shoppingCartItemItem){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             await _shoppingcartService.AddCartItem(User.Identity.Name, shoppingCartItemItem);
             return new ObjectResult("ShoppingCartItem added to database") { StatusCode = StatusCodes.Status201Created };
         }
@@ -36,6 +43,9 @@ namespace Cryptocop.Software.API.Controllers
         [Route("{id}")]
         //Deletes an item from the shopping cart for current user
         public IActionResult RemoveCartItem(int id){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             _shoppingcartService.RemoveCartItem(User.Identity.Name, id);
             return NoContent();
         }
@@ -44,6 +54,9 @@ namespace Cryptocop.Software.API.Controllers
         [Route("{id}")]
         // Updates the quantity for a shopping cart item for current user
         public IActionResult UpdateCartItemQuantity( int id, [FromBody] float quantity){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             _shoppingcartService.UpdateCartItemQuantity(User.Identity.Name, id, quantity);
             return NoContent();
         }
@@ -52,6 +65,9 @@ namespace Cryptocop.Software.API.Controllers
         [Route("")]
         // Clears the cart - all items within the cart should be deleted for current user
         public IActionResult ClearCart(){
+            if (User.Identity == null){
+                throw new IdentityException();
+            }
             _shoppingcartService.ClearCart(User.Identity.Name);
             return NoContent();
         }
